+++ Core/MemMap.cpp
@@ -38,6 +38,7 @@
 #include "Core/Debugger/SymbolMap.h"
 #include "Core/Debugger/Breakpoints.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/HLE/ReplaceTables.h"
 #include "Core/MIPS/JitCommon/JitBlockCache.h"
