+++ Core/MemMapFunctions.cpp
@@ -21,6 +21,7 @@
 #include "Core/Core.h"
 #include "Core/MemMap.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/Host.h"
 #include "Core/Reporting.h"
