+++ Core/PSPLoaders.cpp
@@ -48,6 +48,7 @@
 #include "Host.h"
 
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/System.h"
 #include "Core/PSPLoaders.h"
 #include "Core/HLE/HLE.h"
