+++ GPU/Common/GPUStateUtils.cpp
@@ -22,6 +22,7 @@
 
 #include "Common/StringUtils.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/System.h"
 
 #include "GPU/ge_constants.h"
