+++ GPU/Directx9/GPU_DX9.cpp
@@ -28,6 +28,7 @@
 #include "Core/MIPS/MIPS.h"
 #include "Core/Host.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/Reporting.h"
 #include "Core/System.h"
