+++ GPU/Vulkan/FramebufferVulkan.cpp
@@ -35,6 +35,7 @@
 #include "Core/Host.h"
 #include "Core/MemMap.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/System.h"
 #include "Core/Reporting.h"
 #include "Core/HLE/sceDisplay.h"
