+++ UI/DisplayLayoutScreen.cpp
@@ -26,6 +26,7 @@
 
 #include "DisplayLayoutScreen.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/System.h"
 #include "DisplayLayoutEditor.h"
 #include "GPU/Common/FramebufferCommon.h"
