+++ Windows/DSoundStream.cpp
@@ -12,6 +12,7 @@
 
 #include "thread/threadutil.h"
 #include "Common/OSVersion.h"
+#include "Core/ConfigValues.h"
 #include "Core/Reporting.h"
 #include "Core/Util/AudioFormat.h"
 #include "Windows/W32Util/Misc.h"
