+++ Windows/EmuThread.cpp
@@ -22,6 +22,7 @@
 #include "Core/Host.h"
 #include "Core/System.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "thread/threadutil.h"
 
 enum class EmuThreadState {
