+++ Windows/GPU/D3D9Context.cpp
@@ -8,6 +8,7 @@
 #include "i18n/i18n.h"
 
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/Reporting.h"
 #include "Core/System.h"
 #include "Common/OSVersion.h"
