+++ libretro/LibretroGLContext.cpp
@@ -1,6 +1,7 @@
 
 #include "Common/Log.h"
 #include "Core/Config.h"
+#include "Core/ConfigValues.h"
 #include "Core/System.h"
 #include "gfx_es2/gpu_features.h"
